+++ packages/react/src/hooks/index.ts
@@ -15,6 +15,7 @@ export * from './useSendTransaction.js';
 export * from './useTangledConfig.js';
 export * from './useToken.js';
 export * from './useTokenForAccount.js';
+export * from './useTokenHandlers.js';
 export * from './useTransactionReceipt.js';
 export * from './useTronStore.js';
 export * from './useWaitForTransaction.js';