+++ packages/react/src/hooks/useTokenForAccount.ts
@@ -1,7 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { formatUnits } from 'viem';
-import { getTokenBalanceAndAllowance } from '../actions/getToken.js';
+import { GetTokenBalanceAndAllowanceResponse, getTokenBalanceAndAllowance } from '../actions/getToken.js';
 import { ChainId } from '../types/index.js';
+import { QueryParameter } from '../types/properties.js';
 import { useChain } from './useChain.js';
 import { useConnectionOrConfig } from './useConnectionOrConfig.js';
 import { useToken } from './useToken.js';
@@ -16,10 +17,17 @@ export type UseTokenForAccountParams = {
   /** Allowance spender */
   spender: string | undefined;
   /** Subscribe to token balance and allowance changes for every block */
-  subscribe?: boolean;
+  // subscribe?: boolean;
+  queryOptions?: QueryParameter;
 };
 
-export const useTokenForAccount = ({ chainId, account, token, spender }: UseTokenForAccountParams) => {
+export const useTokenForAccount = ({
+  chainId,
+  account,
+  token,
+  spender,
+  queryOptions = {},
+}: UseTokenForAccountParams) => {
   const chain = useChain(chainId);
   const connectionOrConfig = useConnectionOrConfig();
 
@@ -35,13 +43,33 @@ export const useTokenForAccount = ({ chainId, account, token, spender }: UseToke
         throw new Error('Connections or config not found');
       }
 
-      const { balance, allowance } = await getTokenBalanceAndAllowance({
+      const tokenBalanceAndAllowanceResult = await getTokenBalanceAndAllowance({
         token: token,
         account,
         spender,
         chain,
         config: connectionOrConfig,
       });
+
+      const { balance, allowance } = tokenBalanceAndAllowanceResult;
+
+      if (chain.type === 'solana') {
+        const { associatedTokenAccountAddress, isAtaDeployed } =
+          tokenBalanceAndAllowanceResult as GetTokenBalanceAndAllowanceResponse<'solana'>;
+
+        return {
+          balance: {
+            value: balance as bigint,
+            formatted: formatUnits(balance, tokenMetadata?.decimals),
+          },
+          associatedTokenAccountAddress: associatedTokenAccountAddress,
+          isAtaDeployed: isAtaDeployed,
+          allowance: {
+            value: allowance as bigint,
+            formatted: formatUnits(allowance, tokenMetadata?.decimals),
+          },
+        };
+      }
 
       return {
         balance: {
@@ -57,5 +85,6 @@ export const useTokenForAccount = ({ chainId, account, token, spender }: UseToke
     staleTime: 1000 * 30, // 30 seconds
     refetchOnWindowFocus: true,
     enabled: Boolean(tokenMetadata && account && chain),
+    ...queryOptions,
   });
 };