+++ packages/react/src/types/properties.ts
@@ -1,0 +1,16 @@
+import { DefaultError, QueryKey } from '@tanstack/react-query';
+import { UseQueryParameters } from 'wagmi/query';
+
+export type QueryParameter<
+  queryFnData = unknown,
+  error = DefaultError,
+  data = queryFnData,
+  queryKey extends QueryKey = QueryKey,
+> = {
+  query?:
+    | Omit<
+        UseQueryParameters<queryFnData, error, data, queryKey>,
+        'queryFn' | 'queryHash' | 'queryKey' | 'queryKeyHashFn' | 'throwOnError'
+      >
+    | undefined;
+};