+++ packages/react/src/utils/getDefaultSupportedChains.ts
@@ -4,7 +4,7 @@ import { solana } from '../chains/solana.js';
 import { sui } from '../chains/sui.js';
 import { suiTestnet } from '../chains/sui.testnet.js';
 import { tronMainnet } from '../chains/tron.js';
-import { ChainData, SupportedChainsByType } from '../types/index.js';
+import { EVMChain, OtherChainData, SupportedChainsByType, TronChain } from '../types/index.js';
 
 const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => {
   const supportedChains: SupportedChainsByType = {
@@ -41,11 +41,11 @@ const getDefaultSupportedChains = (testnet?: boolean): SupportedChainsByType => 
       evm.polygonZkEvm,
       evm.scroll,
       evm.zkSync,
-    ] as ChainData<'evm'>[];
-    supportedChains.solana = [solana] as ChainData<'solana'>[];
-    supportedChains.tron = [tronMainnet] as ChainData<'tron'>[];
-    supportedChains.alephZero = [alephZero] as ChainData<'alephZero'>[];
-    supportedChains.sui = [sui, suiTestnet] as ChainData<'sui'>[];
+    ] as EVMChain[];
+    supportedChains.solana = [solana] as OtherChainData<'solana'>[];
+    supportedChains.tron = [tronMainnet] as TronChain[];
+    supportedChains.alephZero = [alephZero] as OtherChainData<'alephZero'>[];
+    supportedChains.sui = [sui, suiTestnet] as OtherChainData<'sui'>[];
   }
 
   return supportedChains;